export `derefs_to_slice` from methods module
